Add explicit template initializations for transfer function
