+++ src/transfer.h
@@ -10,9 +10,7 @@
 
 #pragma once
 
-namespace leopart
-{
-namespace transfer
+namespace leopart::transfer
 {
 /// Return basis values for each particle in function space
 Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>
@@ -41,7 +39,6 @@ void transfer_to_particles(
 /// Evaluate the basis functions at particle positions in a prescribed cell.
 /// Writes results to an Eigen::Matrix \f$q\f$ and Eigen::Vector \f$f\f$ to
 /// compose the l2 problem \f$q q^T = q f\f$.
-
 std::pair<Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>,
           Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>>
 eval_particle_cell_contributions(
@@ -49,140 +46,4 @@ eval_particle_cell_contributions(
     const Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>&
         basis_values,
     int row_offset, int space_dimension, int block_size);
-//----------------------------------------------------------------------------
-template <typename T>
-void transfer_to_function(
-    std::shared_ptr<dolfinx::function::Function<T>> f, const Particles& pax,
-    const Field& field,
-    const Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>&
-        basis_values)
-{
-
-  std::shared_ptr<const dolfinx::mesh::Mesh> mesh = f->function_space()->mesh();
-  const int tdim = mesh->topology().dim();
-  int ncells = mesh->topology().index_map(tdim)->size_local();
-
-  // Get element
-  assert(f->function_space()->element());
-  std::shared_ptr<const dolfinx::fem::FiniteElement> element
-      = f->function_space()->element();
-  assert(element);
-  const int block_size = element->block_size();
-  const int value_size = element->value_size() / block_size;
-  const int space_dimension = element->space_dimension() / block_size;
-  assert(basis_values.cols() == value_size * space_dimension);
-
-  std::shared_ptr<const dolfinx::fem::DofMap> dm
-      = f->function_space()->dofmap();
-
-  // Vector of expansion_coefficients to be set
-  Eigen::Matrix<T, Eigen::Dynamic, 1>& expansion_coefficients = f->x()->array();
-
-  int row_offset = 0;
-  for (int c = 0; c < ncells; ++c)
-  {
-    std::vector<int> cell_particles = pax.cell_particles()[c];
-    auto [q, l] = eval_particle_cell_contributions(cell_particles, field,
-                                                   basis_values, row_offset,
-                                                   space_dimension, block_size);
-
-    // Solve projection where
-    // - q has shape [ndofs/block_size, np]
-    // - l has shape [np, block_size]
-    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> u_tmp
-        = (q * q.transpose()).ldlt().solve(q * l);
-    Eigen::Map<Eigen::VectorXd> u_i(u_tmp.data(), space_dimension * block_size);
-
-    auto dofs = dm->cell_dofs(c);
-
-    assert(dofs.size() == space_dimension);
-
-    for (int i = 0; i < dofs.size(); ++i)
-    {
-      expansion_coefficients[dofs[i]] = u_i[i];
-    }
-    row_offset += cell_particles.size();
-  }
-}
-//----------------------------------------------------------------------------
-template <typename T>
-void transfer_to_particles(
-    Particles& pax, Field& field,
-    std::shared_ptr<const dolfinx::function::Function<T>> f,
-    const Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>&
-        basis_values)
-{
-  // Get element
-  assert(f->function_space()->element());
-  std::shared_ptr<const dolfinx::fem::FiniteElement> element
-      = f->function_space()->element();
-  assert(element);
-  const int block_size = element->block_size();
-  const int value_size = element->value_size() / block_size;
-  const int space_dimension = element->space_dimension() / block_size;
-  assert(basis_values.cols() == value_size * space_dimension);
-  assert(field.value_size() == value_size);
-
-  std::shared_ptr<const dolfinx::mesh::Mesh> mesh = f->function_space()->mesh();
-  const int tdim = mesh->topology().dim();
-  int ncells = mesh->topology().index_map(tdim)->size_local();
-
-  // Get particles in each cell
-  const std::vector<std::vector<int>>& cell_particles = pax.cell_particles();
-
-  // Count up particles in each cell
-  std::shared_ptr<const dolfinx::fem::DofMap> dm
-      = f->function_space()->dofmap();
-
-  // Const array of expansion coefficients
-  const Eigen::Matrix<T, Eigen::Dynamic, 1>& f_array = f->x()->array();
-
-  int idx = 0;
-  for (int c = 0; c < ncells; ++c)
-  {
-    auto dofs = dm->cell_dofs(c);
-    Eigen::VectorXd vals(dofs.size());
-    for (int k = 0; k < dofs.size(); ++k)
-    {
-      vals[k] = f_array[dofs[k]];
-    }
-    // Cast as matrix of size [block_size, space_dimension/block_size]
-    Eigen::Map<const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic,
-                                   Eigen::ColMajor>>
-        vals_mat(vals.data(), block_size, space_dimension);
-
-    for (int pidx : cell_particles[c])
-    {
-      Eigen::Map<Eigen::VectorXd> ptr = field.data(pidx);
-      ptr.setZero();
-      Eigen::Map<const Eigen::VectorXd> q(basis_values.row(idx++).data(),
-                                          space_dimension);
-
-      ptr = vals_mat * q;
-    }
-  }
-}
-//----------------------------------------------------------------------------
-std::pair<Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>,
-          Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>>
-eval_particle_cell_contributions(
-    const std::vector<int>& cell_particles, const Field& field,
-    const Eigen::Array<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>&
-        basis_values,
-    int row_offset, int space_dimension, int block_size)
-{
-  int np = cell_particles.size();
-  Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> q(space_dimension, np),
-      l(np, block_size);
-  for (int p = 0; p < np; ++p)
-  {
-    int pidx = cell_particles[p];
-    Eigen::Map<const Eigen::VectorXd> basis(
-        basis_values.row(row_offset++).data(), space_dimension);
-    q.col(p) = basis;
-    l.row(p) = field.data(pidx);
-  }
-  return std::make_pair(q, l);
-}
-} // namespace transfer
-} // namespace leopart
+} // namespace leopart::transfer
